Code optimizations
Moved variable definitions for save path and date string to top of
script.

Removed GetDate function and converted to using built in diatomite
string format method.

Updated home path to use ~ instead of environment variable.  This makes
it more OS agnostic.

Included SecureCRT IgnoreEscape in the screen capture so as to skip
color escape sequences in the capture.